Alter test_runner to be able to run scripts using custom settings.
---------------- Main.py

This is achieved by reading the optional "--settings" parameter when calling the script.  This parameter points to the location of the "settings.json" file.  When this parameter is not set the program will look for the settings file in (previously always loaded in C:/ProgramData/TingusData/save_files)

The settings file contains the following info:

{
    "testSettings": {
        "runTestDelay": 2,
        "actionDelayOffset": 0,
        "logName": "c:/aaa/tingus_runner.log",
        "save_folder": "F:/github/goodxtests/desktop_app_tests/TingusData/save_files"
    }
}

------------ [ runner.py ]
read the location of the test file from the "settings.json" file